Switch to one_call method
This way we only have to make one api call instead of three. 
We will lose the 3 hour forecast for 5 days, but I think a simple daily forecast for 7 days, is more then enough for most users. 
There is still be an hourly forecast for the next 48 hours included.   
We will lose